feat(line): init typescript migration
