+++ packages/line/package.json
@@ -45,7 +45,6 @@
     "@nivo/scales": "^0.62.0",
     "@nivo/tooltip": "^0.62.0",
     "@nivo/voronoi": "^0.62.0",
-    "prop-types": ">= 15.5.10 < 16.0.0",
     "react": ">= 16.8.4 < 17.0.0"
   },
   "publishConfig": {
