+++ packages/line/src/ResponsiveLineCanvas.tsx
@@ -1,0 +1,24 @@
+/*
+ * This file is part of the nivo project.
+ *
+ * Copyright 2016-present, Raphaël Benitte.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+import React from 'react'
+import { ResponsiveWrapper } from '@nivo/core'
+import LineCanvas from './LineCanvas'
+import { LineBaseProps } from './hooks'
+
+export interface ResponsiveLineCanvasProps extends LineBaseProps {
+    pixelRatio?: number
+}
+
+export default function ResponsiveLineCanvas(props: ResponsiveLineCanvasProps) {
+    return (
+        <ResponsiveWrapper>
+            {({ width, height }) => <LineCanvas width={width} height={height} {...props} />}
+        </ResponsiveWrapper>
+    )
+}
