+++ packages/line/src/hooks.ts
@@ -7,36 +7,217 @@
  * file that was distributed with this source code.
  */
 import { useMemo } from 'react'
-import { area, line } from 'd3-shape'
-import { curveFromProp, useTheme, useValueFormatter } from '@nivo/core'
-import { useOrdinalColorScale, useInheritedColor } from '@nivo/colors'
-import { computeXYScalesForSeries } from '@nivo/scales'
+import { area, line, CurveFactory } from 'd3-shape'
+import { 
+  curveFromProp,
+  useTheme,
+  useValueFormatter,
+  Box,
+  Theme,
+  DataFormatter,
+  DatumValue as CoreDatumValue,
+  CartesianMarkerSpec,
+  CurveInterpolationId,
+  Dimensions
+} from '@nivo/core'
+import { LegendProps } from '@nivo/legends'
+import { AxisProps, GridValues } from '@nivo/axes'
+import { CrosshairType } from '@nivo/tooltip'
+import { useInheritedColor, useOrdinalColorScale, OrdinalColorScale, InheritedColorGenerator } from '@nivo/colors'
+import { computeXYScalesForSeries, ScaleOptions, Scale } from '@nivo/scales'
 import { LineDefaultProps } from './props'
+import { ResponsiveLineProps } from './ResponsiveLine'
+import { SliceTooltipProps } from './SliceTooltip'
+import { PointTooltipProps } from './PointTooltip'
+import { LineProps } from './Line'
 
-export const useLineGenerator = ({ curve }) => {
-    return useMemo(
-        () =>
-            line()
-                .defined(d => d.x !== null && d.y !== null)
-                .x(d => d.x)
-                .y(d => d.y)
-                .curve(curveFromProp(curve)),
-        [curve]
-    )
+export type DatumValue = CoreDatumValue
+
+export interface Datum {
+  x: DatumValue | null
+  y: DatumValue | null
+  [key: string]: any
+}
+export interface ComputedDatum {
+  data: Datum
+  position: {
+    x: number
+    y: number
+  }
 }
 
-export const useAreaGenerator = ({ curve, yScale, areaBaselineValue }) => {
-    return useMemo(() => {
-        return area()
-            .defined(d => d.x !== null && d.y !== null)
-            .x(d => d.x)
-            .y1(d => d.y)
-            .curve(curveFromProp(curve))
-            .y0(yScale(areaBaselineValue))
-    }, [curve, yScale, areaBaselineValue])
+export interface Serie {
+  id: string | number
+  data: Datum[]
+  [key: string]: any
+}
+export interface ComputedSerie {
+  id: string | number
+  data: ComputedDatum[]
+  color?: string
+  [key: string]: any
 }
 
-const usePoints = ({ series, getPointColor, getPointBorderColor, formatX, formatY }) => {
+export type LineLayerType =
+  | 'grid'
+  | 'markers'
+  | 'axes'
+  | 'areas'
+  | 'crosshair'
+  | 'lines'
+  | 'slices'
+  | 'points'
+  | 'mesh'
+  | 'legends'
+
+export type LineGenerator = (data: Datum[]) => string | null
+
+export type Curve = CurveInterpolationId
+
+export interface CustomLayerProps extends Omit<ResponsiveLineProps, 'xScale' | 'yScale'> {
+  innerHeight: number
+  innerWidth: number
+  lineGenerator: LineGenerator
+  points: Point[]
+  series: ComputedSerie[]
+  xScale: Scale
+  yScale: Scale
+}
+
+export type CustomLayer = (props: CustomLayerProps) => React.ReactNode
+export type Layer = LineLayerType | CustomLayer
+
+export interface Point {
+  id: string
+  index: number
+  serieId: string | number
+  serieColor: string
+  x: number
+  y: number
+  color: string
+  borderColor: string
+  data: {
+    x: DatumValue
+    xFormatted: string | number
+    y: DatumValue
+    yFormatted: string | number
+    yStacked?: number
+  }
+}
+
+export type AccessorFunc = (datum: Point['data']) => string
+
+export type PointMouseHandler = (point: Point, event: React.MouseEvent) => void
+
+export interface Slice {
+  id: DatumValue
+  height: number
+  width: number
+  x0: number
+  x: number
+  y0: number
+  y: number
+  points: Point[]
+}
+
+export interface PointSymbolProps {
+  borderColor: string
+  borderWidth: number
+  color: string
+  datum: Datum
+  size: number
+}
+
+export type PointSymbol = React.ComponentClass<PointSymbolProps>
+
+export interface LineBaseProps {
+  data: Serie[]
+  xScale?: ScaleOptions
+  xFormat?: string | DataFormatter
+  yScale?: ScaleOptions
+  yFormat?: string | DataFormatter
+  layers?: Layer[]
+  margin?: Box
+  curve?: Curve
+  lineWidth?: number
+  colors?: OrdinalColorScale<ComputedSerie>
+  theme?: Theme
+  axisTop?: AxisProps
+  axisRight?: AxisProps
+  axisBottom?: AxisProps
+  axisLeft?: AxisProps
+  enableGridX?: boolean
+  gridXValues?: GridValues<DatumValue>
+  enableGridY?: boolean
+  gridYValues?: GridValues<DatumValue>
+  enablePoints?: boolean
+  pointSymbol?: PointSymbol
+  pointSize?: number
+  pointColor?: any
+  pointBorderWidth?: number
+  pointBorderColor?: any
+  enableArea?: boolean
+  areaOpacity?: number
+  areaBaselineValue?: DatumValue
+  markers?: CartesianMarkerSpec<number>[]
+  isInteractive?: boolean
+  onMouseEnter?: PointMouseHandler
+  onMouseMove?: PointMouseHandler
+  onMouseLeave?: PointMouseHandler
+  onClick?: PointMouseHandler
+  debugMesh?: boolean
+  enableSlices?: 'x' | 'y' | false
+  debugSlices?: boolean
+  sliceTooltip?: React.ComponentClass<SliceTooltipProps>
+  tooltipFormat?: DataFormatter | string
+  tooltip?: React.ComponentClass<PointTooltipProps>
+  enableCrosshair?: boolean
+  crosshairType?: CrosshairType
+  legends?: LegendProps[]
+}
+
+export const useLineGenerator = ({
+  curve
+}: {
+  curve: Curve
+}): LineGenerator => {
+  return useMemo(
+    () =>
+      line<any>()
+        .defined((d) => d.x !== null && d.y !== null)
+        .x((d) => d.x)
+        .y((d) => d.y)
+        .curve(curveFromProp(curve)),
+    [curve]
+  )
+}
+
+const useAreaGenerator = ({
+  curve,
+  yScale,
+  areaBaselineValue
+}: {
+  curve: Curve
+  yScale: Scale
+  areaBaselineValue: DatumValue
+}) => {
+  return useMemo(() => {
+    return area<any>()
+      .defined((d) => d.x !== null && d.y !== null)
+      .x((d) => d.x)
+      .y1((d) => d.y)
+      .curve(curveFromProp(curve) as CurveFactory)
+      .y0(yScale(areaBaselineValue))
+  }, [curve, yScale, areaBaselineValue])
+}
+
+const usePoints = ({ series, getPointColor, getPointBorderColor, formatX, formatY }: {
+  series: ComputedSerie[],
+  getPointColor: InheritedColorGenerator<any>,
+  getPointBorderColor: InheritedColorGenerator<any>,
+  formatX: any,
+  formatY: any
+}): Point[] => {
     return useMemo(() => {
         return series.reduce((acc, serie) => {
             return [
@@ -44,7 +225,7 @@ const usePoints = ({ series, getPointColor, getPointBorderColor, formatX, format
                 ...serie.data
                     .filter(datum => datum.position.x !== null && datum.position.y !== null)
                     .map((datum, i) => {
-                        const point = {
+                        const point: Partial<Point> = {
                             id: `${serie.id}.${i}`,
                             index: acc.length + i,
                             serieId: serie.id,
@@ -67,7 +248,17 @@ const usePoints = ({ series, getPointColor, getPointBorderColor, formatX, format
     }, [series, getPointColor, getPointBorderColor, formatX, formatY])
 }
 
-export const useSlices = ({ enableSlices, points, width, height }) => {
+interface UseSlicesProps extends Dimensions {
+  enableSlices: 'x' | 'y' | false
+  points: Point[]
+}
+
+export const useSlices = ({
+  enableSlices,
+  points,
+  width,
+  height
+}: UseSlicesProps): Slice[] => {
     return useMemo(() => {
         if (enableSlices === false) return []
 
@@ -152,8 +343,8 @@ export const useLine = ({
     areaBaselineValue = LineDefaultProps.areaBaselineValue,
     pointColor = LineDefaultProps.pointColor,
     pointBorderColor = LineDefaultProps.pointBorderColor,
-    enableSlices = LineDefaultProps.enableSlicesTooltip,
-}) => {
+    enableSlices = LineDefaultProps.enableSlices,
+}: Pick<LineProps, 'data' | 'xScale' | 'xFormat' | 'yScale' | 'yFormat' | 'colors' | 'width' | 'height' | 'curve' | 'areaBaselineValue' | 'pointColor' | 'pointBorderColor' | 'enableSlices'>) => {
     const formatX = useValueFormatter(xFormat)
     const formatY = useValueFormatter(yFormat)
     const getColor = useOrdinalColorScale(colors, 'id')
@@ -161,12 +352,16 @@ export const useLine = ({
     const getPointColor = useInheritedColor(pointColor, theme)
     const getPointBorderColor = useInheritedColor(pointBorderColor, theme)
 
-    const { xScale, yScale, series: rawSeries } = useMemo(
+    const { xScale, yScale, series: rawSeries }: {
+      xScale: Scale,
+      yScale: Scale,
+      series: ComputedSerie[]
+    } = useMemo(
         () => computeXYScalesForSeries(data, xScaleSpec, yScaleSpec, width, height),
         [data, xScaleSpec, yScaleSpec, width, height]
     )
 
-    const series = useMemo(
+    const series: ComputedSerie[] = useMemo(
         () =>
             rawSeries.map(serie => ({
                 ...serie,
